+++ examples/react-native/WhipWhepDemo/app/(tabs)/index.tsx
@@ -1,63 +1,17 @@
-import { StyleSheet, Button, View, ActivityIndicator } from 'react-native';
-
-import React, { useEffect, useState } from 'react';
-import {
-  connectWhepClient,
-  createWhepClient,
-  disconnectWhepClient,
-  pauseWhepClient,
-  unpauseWhepClient,
-  WhepClientView,
-} from 'react-native-whip-whep';
-import { checkPermissions } from '@/utils/CheckPermissions';
+import { Button, View, ActivityIndicator } from 'react-native';
+import { useWhepClient } from '@/hooks/useWhepClient';
+import { WhepClientView } from 'react-native-whip-whep';
+import { styles } from '../../styles/styles';
 
 export default function HomeScreen() {
-  const [isLoading, setIsLoading] = useState(false);
-  const [shouldShowPlayBtn, setShouldShowPlayBtn] = useState(true);
-  const [isPaused, setIsPaused] = useState(false);
-
-  const handlePlayBtnClick = async () => {
-    setShouldShowPlayBtn(false);
-    setIsLoading(true);
-    try {
-      await connectWhepClient();
-      setIsLoading(false);
-    } catch (error) {
-      console.error('Failed to connect to WHEP Client', error);
-    }
-  };
-
-  const handlePauseBtnClick = async () => {
-    try {
-      pauseWhepClient();
-      setIsPaused(true);
-    } catch (error) {
-      console.error('Failed to pause WHEP Client', error);
-    }
-  };
-
-  const handleRestartBtnClick = async () => {
-    try {
-      unpauseWhepClient();
-      setIsPaused(false);
-    } catch (error) {
-      console.error('Failed to unpause WHEP Client', error);
-    }
-  };
-
-  const initialize = async () => {
-    await checkPermissions();
-    createWhepClient('https://broadcaster.elixir-webrtc.org/api/whep', {
-      authToken: 'example',
-    });
-  };
-
-  useEffect(() => {
-    initialize();
-    return () => {
-      disconnectWhepClient();
-    };
-  }, []);
+  const {
+    isLoading,
+    shouldShowPlayBtn,
+    isPaused,
+    handlePlayBtnClick,
+    handlePauseBtnClick,
+    handleRestartBtnClick,
+  } = useWhepClient('https://broadcaster.elixir-webrtc.org/api/whep');
 
   return (
     <View style={styles.container}>
@@ -78,20 +32,3 @@ export default function HomeScreen() {
     </View>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    paddingTop: 50,
-  },
-  box: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  clientView: {
-    width: '100%',
-    height: 200,
-    marginBottom: 20,
-  },
-});
