+++ examples/react-native/WhipWhepDemo/styles/styles.js
@@ -1,0 +1,18 @@
+import { StyleSheet } from 'react-native';
+
+export const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    paddingTop: 50,
+  },
+  box: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+  clientView: {
+    width: '100%',
+    height: 200,
+    marginBottom: 20,
+  },
+});
